third Push

+++ ft_memset.c
@@ -6,14 +6,16 @@
 /*   By: uclement <uclement@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 11:38:22 by uclement          #+#    #+#             */
-/*   Updated: 2022/11/14 13:23:41 by uclement         ###   ########.fr       */
+/*   Updated: 2022/11/16 15:02:57 by uclement         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-char	*ft_memset(void *str, int c, int len)
+#include <stddef.h>
+
+char	*ft_memset(void *str, int c, size_t len)
 {
-	int		i;
-	char	*str1;
+	size_t		i;
+	char		*str1;
 
 	str1 = ((char *) str);
